refactor: 詳細情報をコンポーネントに切り出した
